requirements.txt
# requirements.txt
# This file specifies which Python dependencies are required by this project.
#
# To install them, navigate the terminal to the project root ( lf6/ ) and run
# the following command: pip install -r requirements.txt

# PyYAML is used for deserializing YAML-based config files
pyyaml >= 6.0.1

# OpenAI is used to query the fine-tuned openai model backing the chatbot
openai >= 1.5.0

# Flask is used for serving the endpoints and frontend of the chatbot
flask >= 3.0.0

# Pandas is used to implement the classification model
pandas >= 2.1.4

# Scikit (sklearn) is used to implement the classification model
scikit-learn >= 1.3.2
